Refactored yelp to separate component
